Add tests for safety fixes in AIDL generation
- Implement tests to handle circular reference detection and infinite loop prevention.
- Validate graceful handling of unresolvable references and deep nesting stability.
- Ensure performance with large enums and mixed resolvable/unresolvable references.
- Include a simplified test for Android KeyMint style enums.
